User and Roles comments fixed
