+++ plugins/modules/user_role_workflow_manager.py
@@ -848,13 +848,6 @@ response_11:
     }
 """
 
-try:
-    from cryptography.fernet import Fernet
-    HAS_FERNET = True
-except ImportError:
-    HAS_FERNET = False
-    Fernet = None
-
 import re
 from ansible_collections.cisco.dnac.plugins.module_utils.dnac import (
     DnacBase,
@@ -897,12 +890,10 @@ class UserandRole(DnacBase):
         """
         self.log("Validating the Playbook Yaml File..", "INFO")
         config = self.payload.get("config")
-        self.key = None
+        self.key = self.generate_key()
 
-        if HAS_FERNET:
-            self.key = self.generate_key()
-        else:
-            self.msg = "The 'cryptography' library is not installed. Please install it using 'pip install cryptography'."
+        if self.key and "error_message" in self.key:
+            self.msg = self.key.get("error_message")
             self.log(self.msg, "ERROR")
             self.status = "failed"
             return self
@@ -944,7 +935,15 @@ class UserandRole(DnacBase):
         if "user_details" in config and "username" in user_role_details[0] or "email" in user_role_details[0]:
             for user in user_role_details:
                 if 'password' in user:
-                    user['password'] = self.encrypt_password(user['password'], self.key)
+                    encrypt_password_responce = self.encrypt_password(user['password'], self.key.get("generate_key"))
+
+                    if encrypt_password_responce and "error_message" in encrypt_password_responce:
+                        self.msg = encrypt_password_responce.get("error_message")
+                        self.log(self.msg, "ERROR")
+                        self.status = "failed"
+                        return self
+
+                    user["password"] = encrypt_password_responce.get("encrypt_password")
 
             if user_role_details[0].get("username") is not None or user_role_details[0].get("email") is not None:
                 user_details = {
@@ -992,48 +991,6 @@ class UserandRole(DnacBase):
         """
         if field_value and not regex.match(field_value):
             error_messages.append(error_message)
-
-    def generate_key(self):
-        """
-        Generate a new encryption key using Fernet.
-        Returns:
-            - key (bytes): A newly generated encryption key.
-        Criteria:
-            - This function should only be called if HAS_FERNET is True.
-        """
-        return Fernet.generate_key()
-
-    def encrypt_password(self, password, key):
-        """
-        Encrypt a plaintext password using the provided encryption key.
-        Args:
-            - password (str): The plaintext password to be encrypted.
-            - key (bytes): The encryption key used to encrypt the password.
-        Returns:
-            - encrypted_password (bytes): The encrypted password as bytes.
-        Criteria:
-            - This function should only be called if HAS_FERNET is True.
-            - The password should be encoded to bytes before encryption.
-        """
-        fernet = Fernet(key)
-        encrypted_password = fernet.encrypt(password.encode())
-        return encrypted_password
-
-    def decrypt_password(self, encrypted_password, key):
-        """
-        Decrypt an encrypted password using the provided encryption key.
-        Args:
-            - encrypted_password (bytes): The encrypted password as bytes to be decrypted.
-            - key (bytes): The encryption key used to decrypt the password.
-        Returns:
-            - decrypted_password (str): The decrypted plaintext password.
-        Criteria:
-            - This function should only be called if HAS_FERNET is True.
-            - The encrypted password should be decoded from bytes after decryption.
-        """
-        fernet = Fernet(key)
-        decrypted_password = fernet.decrypt(encrypted_password.encode()).decode()
-        return decrypted_password
 
     def validate_password(self, password, error_messages):
         """
@@ -1240,16 +1197,28 @@ class UserandRole(DnacBase):
 
         password = user_config.get("password")
 
-        try:
-            if password:
-                user_config['password'] = self.decrypt_password(password, self.key)
-                plain_password = user_config.get("password")
-                self.validate_password(plain_password, error_messages)
-                user_config['password'] = self.encrypt_password(plain_password, self.key).decode()
-                self.log("Password decrypted, validated, and re-encrypted successfully.", "DEBUG")
-        except Exception as e:
-            self.msg = "Failed during password processing: {0}".format(str(e))
-            self.log(self.msg, "ERROR")
+        if password:
+            decrypt_password_responce = self.decrypt_password(password, self.key.get("generate_key"))
+
+            if decrypt_password_responce and "error_message" in decrypt_password_responce:
+                self.msg = decrypt_password_responce.get("error_message")
+                self.log(self.msg, "ERROR")
+                self.status = "failed"
+                return self
+
+            user_config['password'] = decrypt_password_responce.get("decrypt_password")
+            plain_password = user_config.get("password")
+            self.validate_password(plain_password, error_messages)
+            encrypt_password_responce = self.encrypt_password(plain_password, self.key.get("generate_key"))
+
+            if encrypt_password_responce and "error_message" in encrypt_password_responce:
+                self.msg = encrypt_password_responce.get("error_message")
+                self.log(self.msg, "ERROR")
+                self.status = "failed"
+                return self
+
+            user_config['password'] = encrypt_password_responce.get("encrypt_password").decode()
+            self.log("Password decrypted, validated, and re-encrypted successfully.", "DEBUG")
 
         username_regex = re.compile(r"^[A-Za-z0-9@._-]{3,50}$")
         username_regex_msg = "The username must not contain any special characters and must be 3 to 50 characters long."
@@ -1575,7 +1544,15 @@ class UserandRole(DnacBase):
             - Returns the API response from the "create_user" function.
         """
         self.log("Create user with 'user_params' argument...", "DEBUG")
-        user_params['password'] = self.decrypt_password(user_params['password'], self.key)
+        decrypt_password_responce = self.decrypt_password(user_params['password'], self.key.get("generate_key"))
+
+        if decrypt_password_responce and "error_message" in decrypt_password_responce:
+            self.msg = decrypt_password_responce.get("error_message")
+            self.log(self.msg, "ERROR")
+            self.status = "failed"
+            return self
+
+        user_params['password'] = decrypt_password_responce.get("decrypt_password")
         required_keys = ['username', 'password']
         missing_keys = []
